cleaning up some div by 0 warnings for not full 3 phase systems
