+++ ic.txt
@@ -1,10 +1,11 @@
 v(x,0)
 v(w,100)
 v(h,10)
-v(p,2)
+v(p,0)
 
 
 l(1)
+a(p,1)
 a(x,150)
 v(y,0)
